Move the setup validators to notices
