+++ app/services/notices/setup/submit-remove-licences.service.js
@@ -7,7 +7,7 @@
 
 const FetchReturnsDueService = require('./fetch-returns-due.service.js')
 const RemoveLicencesPresenter = require('../../../presenters/notifications/setup/remove-licences.presenter.js')
-const RemoveLicencesValidator = require('../../../validators/notifications/setup/remove-licences.validator.js')
+const RemoveLicencesValidator = require('../../../validators/notices/setup/remove-licences.validator.js')
 const SessionModel = require('../../../models/session.model.js')
 const { transformStringOfLicencesToArray } = require('../../../lib/general.lib.js')
