+++ app/validators/notices/setup/remove-licences.validator.js
